Refactor some tests
