+++ test/variant.test.ts
@@ -1,4 +1,4 @@
-import data from './data';
+import data from '../data/variants';
 import {
     stringifyVariant,
     stripParentheses,
